src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/primitive/ArrayPrimitiveIntegerAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.primitive;

import com.github.brunomndantas.jscrapper.support.parser.array.attribute.ArrayAttributeParser;
import org.openqa.selenium.WebElement;

public class ArrayPrimitiveIntegerAttributeParser extends ArrayAttributeParser {

    public ArrayPrimitiveIntegerAttributeParser(String attribute) {
        super(int.class, attribute, 0);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Integer.parseInt(element.getAttribute(super.attribute));
    }

}
