src/test/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/primitive/ArrayPrimitiveIntegerAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.primitive;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class ArrayPrimitiveIntegerAttributeParserTest {

    @Test
    public void constructorTest() {
        String attribute = "";
        ArrayPrimitiveIntegerAttributeParser parser = new ArrayPrimitiveIntegerAttributeParser(attribute);
        assertSame(int.class, parser.getKlass());
        assertEquals(0, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        ArrayPrimitiveIntegerAttributeParser parser = new ArrayPrimitiveIntegerAttributeParser("");

        assertEquals(2, parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "2";
            }
        }));
    }

}
